[DEV-6108] Split controllers.goal_controller module
