+++ server/athenian/api/controllers/goal_controller/__init__.py
@@ -1,0 +1,9 @@
+from athenian.api.controllers.goal_controller.goal import create_goal, delete_goal, update_goal
+from athenian.api.controllers.goal_controller.goal_template import (
+    create_goal_template,
+    delete_goal_template,
+    get_goal_template,
+    list_goal_templates,
+    update_goal_template,
+)
+from athenian.api.controllers.goal_controller.measure_goals import measure_goals
